rollback changes to deserializer, move away from custom request composer and use interceptor, add extension function to apply both interceptors at once
