+++ sentry-apollo-3/src/main/java/io/sentry/apollo3/apolloBuilderExtensions.kt
@@ -1,0 +1,12 @@
+package io.sentry.apollo3
+
+import com.apollographql.apollo3.ApolloClient
+import io.sentry.HubAdapter
+import io.sentry.IHub
+
+@JvmOverloads
+fun ApolloClient.Builder.sentryTracing(hub: IHub = HubAdapter.getInstance(), beforeSpan: SentryApollo3HttpInterceptor.BeforeSpanCallback? = null): ApolloClient.Builder {
+    addInterceptor(SentryApollo3Interceptor())
+    addHttpInterceptor(SentryApollo3HttpInterceptor(hub, beforeSpan))
+    return this
+}
